Add login and logout to links of sidebar
