+++ src/Components/Sidebar.js
@@ -1,15 +1,11 @@
 import React, { useState } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
 import { FaBars, FaTimes } from 'react-icons/fa';
 import Navlinks from './Navlinks';
 import social from '../Utils/social';
 import Logo from '../Utils/glamping.png';
-import { logout } from '../store/actions/userActions';
 
 const Sidebar = () => {
   const [sidebarOpen, setSidebarOpen] = useState(true);
-  const username = useSelector((state) => state.user.username);
-  const dispatch = useDispatch();
 
   const handleSidebarOpen = () => {
     setSidebarOpen(true);
@@ -18,12 +14,6 @@ const Sidebar = () => {
   const handleSidebarClose = () => {
     setSidebarOpen(false);
   };
-
-  const handleLogout = () => {
-    dispatch(logout());
-    window.location.reload();
-  };
-
   return (
     <div className={`sidebar ${sidebarOpen ? 'open-sidebar' : 'close-sidebar'}`}>
       {!sidebarOpen ? (
@@ -44,14 +34,7 @@ const Sidebar = () => {
                 className="logo-img"
               />
             </div>
-
             <Navlinks />
-            {username !== 'guest' && (
-              <button type="button" onClick={handleLogout} className="logout-button">
-                Logout
-              </button>
-            )}
-
             <div className="sidebar-footer">
               <div className="social-icon">
                 {social.map((socialIcon) => {
